sepet bilgileri ve toast yapısı sisteme eklendi

+++ Core/GreenDelight.Domain/Concrete/OrderItem.cs
@@ -11,7 +11,7 @@ namespace GreenDelight.Domain.Concrete
     {
         public int OrderID { get; set; }
         public int ProductID { get; set; }
-        public int Quantity { get; set; }
+        public short Quantity { get; set; }
         public decimal UnitPrice { get; set; }
 
         public virtual Order Order { get; set; }
